+++ CHANGES.rst
@@ -16,7 +16,7 @@
 * New InlineListWidget to list radio options inline
 * Added ParsleyJS validation and reorganized forms to be more modular
 * Added app-global CSRF protection (optional) and removed Mustache dependency
-
+* Added ``firstline`` filter to extract first line of text from a HTML block
 
 0.3.0
 -----